Extract Swiper resizing callback to a separate file, share the file between editor and view. Move init functionality into resize so there is only one callback.
