+++ client/gutenberg/extensions/slideshow/slideshow.js
@@ -8,8 +8,7 @@ import { isEqual } from 'lodash';
  * Internal dependencies
  */
 import createSwiper from './create-swiper';
-
-const SIXTEEN_BY_NINE = 16 / 9;
+import swiperResize from './swiper-resize';
 
 class Slideshow extends Component {
 	static defaultProps = {
@@ -97,23 +96,6 @@ class Slideshow extends Component {
 		);
 	}
 
-	swiperInit = swiper => {
-		swiper.el.classList.add( 'wp-swiper-initialized' );
-		this.swiperAutoSize( swiper );
-	};
-
-	swiperAutoSize = swiper => {
-		const img = swiper.el.querySelector( '.swiper-slide[data-swiper-slide-index="0"] img' );
-		if ( ! img ) {
-			return;
-		}
-		const aspectRatio = img.clientWidth / img.clientHeight;
-		const sanityAspectRatio = Math.max( Math.min( aspectRatio, SIXTEEN_BY_NINE ), 1 );
-		const sanityHeight = typeof window !== 'undefined' ? window.innerHeight * 0.8 : 600;
-		const swiperHeight = Math.min( swiper.width / sanityAspectRatio, sanityHeight );
-		swiper.el.style.height = `calc( ${ Math.floor( swiperHeight ) }px + 4em )`;
-	};
-
 	buildSwiper = ( initialSlide = 0 ) =>
 		// Using refs instead of className-based selectors allows us to
 		// have multiple swipers on one page without collisions, and
@@ -140,9 +122,9 @@ class Slideshow extends Component {
 				},
 			},
 			{
-				init: this.swiperInit,
-				imagesReady: this.swiperAutoSize,
-				resize: this.swiperAutoSize,
+				init: swiperResize,
+				imagesReady: swiperResize,
+				resize: swiperResize,
 			}
 		);
 }
