+++ client/gutenberg/extensions/slideshow/view.js
@@ -7,6 +7,7 @@ import { forEach } from 'lodash';
  * Internal dependencies
  */
 import createSwiper from './create-swiper';
+import swiperResize from './swiper-resize';
 
 typeof window !== 'undefined' &&
 	window.addEventListener( 'load', function() {
@@ -14,16 +15,24 @@ typeof window !== 'undefined' &&
 		forEach( slideshowBlocks, slideshowBlock => {
 			const { autoplay, delay, effect } = slideshowBlock.dataset;
 			const slideshowContainer = slideshowBlock.getElementsByClassName( 'swiper-container' )[ 0 ];
-			createSwiper( slideshowContainer, {
-				autoplay: autoplay ? { delay: delay * 1000 } : false,
-				effect,
-				init: true,
-				initialSlide: 0,
-				loop: true,
-				keyboard: {
-					enabled: true,
-					onlyInViewport: true,
+			createSwiper(
+				slideshowContainer,
+				{
+					autoplay: autoplay ? { delay: delay * 1000 } : false,
+					effect,
+					init: true,
+					initialSlide: 0,
+					loop: true,
+					keyboard: {
+						enabled: true,
+						onlyInViewport: true,
+					},
 				},
-			} );
+				{
+					init: swiperResize,
+					imagesReady: swiperResize,
+					resize: swiperResize,
+				}
+			);
 		} );
 	} );
